+++ backend/jobplatform/src/main/java/jobplatform/fo/user/service/M_MypageService.java
@@ -32,7 +32,7 @@ public interface M_MypageService {
 
 	//@@예정@@포지션 제안 수락했을 때 -> 같은 공고로 지원한 내역 있는지 확인 + 지원 프로세스 진행 + 포지션 제안 상태 변경 
 	//@@예정@@포지션 제안 거절했을 때 -> 포지션 제안 상태 변경
-	public int refuseProposedPostion(int pstn_prpsl_sq);
+	public int refuseProposedPostion(Long pstn_prpsl_sq);
 
 	//회원 정보 받아오기
 	public Map<String, Object> findUserINfo(Long mbr_sq);
